backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
